move sparsity stuff to separate file
