+++ src/eval/eval.h
@@ -6,9 +6,9 @@ using EvalClass = raphael::hce::RaphaelHCE;
 
 #elif defined(EVAL_NNUE)
 #ifdef EVAL_MULTILAYER
-    #include <eval/multilayer.h>
+    #include <eval/nnue_multilayer.h>
 #elif defined(EVAL_SINGLELAYER)
-    #include <eval/singlelayer.h>
+    #include <eval/nnue_singlelayer.h>
 #endif
 
 using EvalClass = raphael::nnue::Nnue;
